Introduce and use the event "enhancecomplete"
The event is triggered on a given element (usually a page) after all its children have been enhanced by a given enhancement handler. The data of the event is a string identifying which enhancements have completed.

This setup can be used in conjunction with a deferred to delay certain enhancements until dependent enhancements have completed.